Feita configuração de app sounds sem postgres
close #127
